move file_writers to directory
