+++ swiftwinrt/file_writers/file_writers.h
