+++ swiftwinrt/types/delegate_type.cpp
@@ -2,7 +2,7 @@
 
 #include "types/delegate_type.h"
 
-#include "abi_writer.h"
+#include "file_writers/abi_writer.h"
 #include "utility/metadata_helpers.h"
 #include "type_writers.h"
 #include "types/class_type.h"
