+++ swiftwinrt/types/element_type.cpp
@@ -2,7 +2,7 @@
 
 #include "types/element_type.h"
 
-#include "abi_writer.h"
+#include "file_writers/abi_writer.h"
 #include "utility/metadata_helpers.h"
 
 using namespace std::literals;
