+++ swiftwinrt/types/enum_type.cpp
@@ -2,7 +2,7 @@
 
 #include "types/enum_type.h"
 
-#include "abi_writer.h"
+#include "file_writers/abi_writer.h"
 #include "utility/metadata_helpers.h"
 #include "types/type_definition_helpers.h"
