+++ swiftwinrt/types/interface_type.cpp
@@ -2,7 +2,7 @@
 
 #include "types/interface_type.h"
 
-#include "abi_writer.h"
+#include "file_writers/abi_writer.h"
 #include "utility/metadata_helpers.h"
 #include "types/class_type.h"
 #include "types/delegate_type.h"
